Registration with PayPal Integration plus major code clean up

+++ root/members/utils/modifyregistration.php
@@ -7,9 +7,9 @@ if (!isset($userSession) || $userSession == "") {
     header("Location: /");
     exit;
 }
-include_once('dbconnect.php');
-include_once('checkadmin.php');
-include_once('check_app_state.php');
+include('dbconnect.php');
+include('checkadmin.php');
+include('check_app_state.php');
 
 function getBooleanValue($val) {
     if (!isset($val) || $val == 0 || $val == "false") {
@@ -22,13 +22,10 @@ function getBooleanValue($val) {
 // Get parameters from the url
 $hasParamAgreeToTerms = isset($_POST['agreeToTerms']);
 $hasParamSetRegistered = isset($_POST['setRegistered']);
-$hasParamSetPaid = isset($_POST['setPaid']);
 $hasParamSetPresent = isset($_POST['setPresent']);
 $hasParamToggleRegistered = isset($_POST['toggleRegistered']);
-$hasParamTogglePaid = isset($_POST['togglePaid']);
 $hasParamTogglePresent = isset($_POST['togglePresent']);
 $setRegistered = getBooleanValue($_POST['setRegistered']);
-$setPaid = getBooleanValue($_POST['setPaid']);
 $setPresent = getBooleanValue($_POST['setPresent']);
 if (isset($_POST['uid'])) {
     $uid = $MySQLi_CON->real_escape_string($_POST['uid']);
@@ -41,15 +38,13 @@ if (!$isAdmin) {
     if ($uid != $userSession) {
         die("Only admins can modify others");
     }
-    $hasParamSetPaid = 0;
     $hasParamSetPresent = 0;
-    $hasParamTogglePaid = 0;
     $hasParamTogglePresent = 0;
 }
 
 // Don't proceed unless we have changes to make
-if ($hasParamAgreeToTerms || $hasParamToggleRegistered || $hasParamTogglePaid || $hasParamTogglePresent || $hasParamSetRegistered || $hasParamSetPaid || $hasParamSetPresent) {
-    $userQuery = "SELECT u.isPresent, u.isRegistered, u.isPaid, u.agreeToTerms
+if ($hasParamAgreeToTerms || $hasParamToggleRegistered || $hasParamTogglePresent || $hasParamSetRegistered || $hasParamSetPresent) {
+    $userQuery = "SELECT u.isPresent, u.isRegistered, u.agreeToTerms
 		 FROM users u
 		 WHERE u.uid = {$uid}";
 } else {
@@ -62,15 +57,11 @@ $user = $result->fetch_array();
 $result->free_result();
 
 // Update the values or fall back to the user's current status
-$isPaid = $hasParamSetPaid ? $setPaid : $user['isPaid'];
 $isPresent = $hasParamSetPresent ? $setPresent : $user['isPresent'];
 $isRegistered = $hasParamSetRegistered ? $setRegistered : $user['isRegistered'];
 $agreeToTerms = $hasParamAgreeToTerms ? date("Y-m-d H:i:s") : $user['agreeToTerms'];
 
 // Toggle the values based on the POST
-if ($hasParamTogglePaid) {
-    $isPaid = ($isPaid ? 0 : 1);
-}
 if ($hasParamTogglePresent) {
     $isPresent = ($isPresent ? 0 : 1);
 }
@@ -83,8 +74,7 @@ $agreeQueryPiece = !!$hasParamAgreeToTerms ? ", u.agreeToTerms = NOW()" : "";
 
 // Update the database with the changes
 $updateQuery = "UPDATE users u
-	 SET u.isPaid = {$isPaid}, u.isPresent = {$isPresent},
-		u.isRegistered = {$isRegistered}{$agreeQueryPiece}
+	 SET u.isPresent = {$isPresent}, u.isRegistered = {$isRegistered}{$agreeQueryPiece}
 	 WHERE u.uid = {$uid}";
 $updateResult = $MySQLi_CON->query($updateQuery);
 if (!$updateResult) {
@@ -97,42 +87,49 @@ $user = $result->fetch_array();
 $result->free_result();
 
 // Check the user's updated status
-$isPaid = $user['isPaid'];
 $isPresent = $user['isPresent'];
 $isRegistered = $user['isRegistered'];
 $agreeToTerms = $user['agreeToTerms'];
-
 
 // Count the `registration_stats` rows for this user that are for this year
 $numRows = 0;
 $checkResult = $MySQLi_CON->query("SELECT * FROM registration_stats s
 	 WHERE s.conYear = {$conYear} AND s.uid = {$uid}");
-while ($checkResult->fetch_array()) {
+while ($result = $checkResult->fetch_array()) {
+    $prevOrderId = $result['orderId'];
     $numRows++;
 }
 $checkResult->free_result();
 
 if ($isRegistered == 0) {
-    // Delete registration stats when users unregister
-    $deleteQuery = "DELETE FROM registration_stats WHERE uid = {$uid} AND conYear = {$conYear}";
-    $MySQLi_CON->query($deleteQuery);
-    $statsOperation = "DELETE";
-    $statsQuery = $deleteQuery;
+    if ($prevOrderId == null) {
+        // Delete registration stats when users unregister (if there's no orderId)
+        $deleteQuery = "DELETE FROM registration_stats WHERE uid = {$uid} AND conYear = {$conYear}";
+        $MySQLi_CON->query($deleteQuery);
+        $statsOperation = "DELETE";
+        $statsQuery = $deleteQuery;
+    } else {
+        // If there is an orderId, just update
+        $updateQuery = "UPDATE registration_stats s
+		 SET s.isRegistered = {$isRegistered}, s.isPresent = {$isPresent}, s.modified = CURRENT_TIMESTAMP()
+		 WHERE s.uid = {$uid} AND s.conYear = {$conYear}";
+        $MySQLi_CON->query($updateQuery);
+        $statsOperation = "UPDATE";
+        $statsQuery = $updateQuery;
+    }
 } else if (!$checkResult || $numRows == 0) {
     // Update the registration stats
 
     // Insert a new row for this year's registration stats for this user
-    $insertQuery = "INSERT INTO `registration_stats`(`uid`, `conYear`, `isPaid`, `isPresent`)
-		VALUES ({$uid}, {$conYear}, {$isPaid}, {$isPresent})";
+    $insertQuery = "INSERT INTO `registration_stats`(`uid`, `conYear`, `isRegistered`, `isPresent`, `orderId`)
+		VALUES ({$uid}, {$conYear}, {$isRegistered}, {$isPresent}, {$orderId})";
     $MySQLi_CON->query($insertQuery);
     $statsOperation = "INSERT";
     $statsQuery = $insertQuery;
 } else {
     // Update this year's registration stats for this user
     $updateQuery = "UPDATE registration_stats s
-		 SET s.isPaid = {$isPaid},
-			s.isPresent = {$isPresent},
-			s.modified = CURRENT_TIMESTAMP()
+		 SET s.isRegistered = {$isRegistered}, s.isPresent = {$isPresent}, s.modified = CURRENT_TIMESTAMP()
 		 WHERE s.uid = {$uid} AND s.conYear = {$conYear}";
     $MySQLi_CON->query($updateQuery);
     $statsOperation = "UPDATE";
@@ -142,7 +139,6 @@ if ($isRegistered == 0) {
 // Build JSON with the updated data
 $str = "{";
 $str = "{$str}\"uid\":\"{$uid}\",";
-$str = "{$str}\"isPaid\":{$isPaid},";
 $str = "{$str}\"isPresent\":{$isPresent},";
 $str = "{$str}\"isRegistered\":{$isRegistered},";
 $str = "{$str}\"agreeToTerms\":\"{$agreeToTerms}\",";
